requirements.txt
python-dotenv 
pandas 
numpy 
pymongo 
pymongo[srv]
certifi
scikit-learn
evidently==0.2.8

##-e .